Use dedicated functions for i18n
A specific call triggers a 'Sessions are disabled for this entry point'
exception in load.php. This fix replaces by an equivalent function.

By the way, it is used here the '''content language''' and not the user
language, which is better else the result could be misinterpreted depending
on the user language.

Change-Id: Icc69115758c2ac0b5d0c87c855187e5d24559caf