Remove bogus array_key_exists check
I checked the Parser, and this can never happen. If it could, this code
would be broken, because it would then skip elements from the end of the
array. See, in a sparse array one or more numeric keys don't exist. count()
skips these and returns a smaller number. Because of the way this for loop
is constructed it would end to early and miss elements from the end of the
array.

This is the relevant line of code that makes sure the array of $args is
*not* sparse:
https://github.com/wikimedia/mediawiki/blob/master/includes/parser/Parser.php#L3182

So we can just remove this.

In case you want to be sure the code works with all kinds of arrays (even
sparse ones) you have two options:
1. Use foreach instead of for.
2. Apply array_values on the array first.

But I figured what this patch does is good enough. Tests are in place to
make sure this does not break anything.

Change-Id: Ib368d8e7569eda1cd983b3427c98bebc316dab19